executor: Migrate dataForStatistics, dataForTikVRegionPeers, dataForTiDBHotRegions from package  to
