+++ executor/infoschema_reader.go
@@ -18,13 +18,18 @@ import (
 	"fmt"
 	"sort"
 
+	"github.com/pingcap/errors"
 	"github.com/pingcap/parser/charset"
 	"github.com/pingcap/parser/model"
 	"github.com/pingcap/parser/mysql"
 	"github.com/pingcap/tidb/infoschema"
 	"github.com/pingcap/tidb/privilege"
 	"github.com/pingcap/tidb/sessionctx"
+	"github.com/pingcap/tidb/store/helper"
+	"github.com/pingcap/tidb/store/tikv"
 	"github.com/pingcap/tidb/types"
+	"github.com/pingcap/tidb/util/pdapi"
+	"github.com/pingcap/tidb/util/set"
 )
 
 type memtableRetriever struct {
@@ -52,6 +57,8 @@ func (e *memtableRetriever) retrieve(ctx context.Context, sctx sessionctx.Contex
 		switch e.table.Name.O {
 		case infoschema.TableSchemata:
 			e.rows = dataForSchemata(sctx, dbs)
+		case infoschema.TableStatistics:
+			e.rows = dataForStatistics(sctx, dbs)
 		case infoschema.TableTiDBIndexes:
 			e.rows, err = dataForIndexes(sctx, dbs)
 		case infoschema.TableViews:
@@ -68,6 +75,10 @@ func (e *memtableRetriever) retrieve(ctx context.Context, sctx sessionctx.Contex
 			e.rows = dataForCollationCharacterSetApplicability()
 		case infoschema.TableUserPrivileges:
 			e.setDataFromUserPrivileges(sctx)
+		case infoschema.TableTiKVRegionPeers:
+			e.rows, err = dataForTikVRegionPeers(sctx)
+		case infoschema.TableTiDBHotRegions:
+			e.rows, err = dataForTiDBHotRegions(sctx)
 		}
 		if err != nil {
 			return nil, err
@@ -129,6 +140,97 @@ func dataForSchemata(ctx sessionctx.Context, schemas []*model.DBInfo) [][]types.
 			nil,
 		)
 		rows = append(rows, record)
+	}
+	return rows
+}
+
+func dataForStatistics(ctx sessionctx.Context, schemas []*model.DBInfo) [][]types.Datum {
+	checker := privilege.GetPrivilegeManager(ctx)
+	var rows [][]types.Datum
+	for _, schema := range schemas {
+		for _, table := range schema.Tables {
+			if checker != nil && !checker.RequestVerification(ctx.GetSessionVars().ActiveRoles, schema.Name.L, table.Name.L, "", mysql.AllPrivMask) {
+				continue
+			}
+
+			rs := dataForStatisticsInTable(schema, table)
+			rows = append(rows, rs...)
+		}
+	}
+	return rows
+}
+
+func dataForStatisticsInTable(schema *model.DBInfo, table *model.TableInfo) [][]types.Datum {
+	var rows [][]types.Datum
+	if table.PKIsHandle {
+		for _, col := range table.Columns {
+			if mysql.HasPriKeyFlag(col.Flag) {
+				record := types.MakeDatums(
+					infoschema.CatalogVal, // TABLE_CATALOG
+					schema.Name.O,         // TABLE_SCHEMA
+					table.Name.O,          // TABLE_NAME
+					"0",                   // NON_UNIQUE
+					schema.Name.O,         // INDEX_SCHEMA
+					"PRIMARY",             // INDEX_NAME
+					1,                     // SEQ_IN_INDEX
+					col.Name.O,            // COLUMN_NAME
+					"A",                   // COLLATION
+					0,                     // CARDINALITY
+					nil,                   // SUB_PART
+					nil,                   // PACKED
+					"",                    // NULLABLE
+					"BTREE",               // INDEX_TYPE
+					"",                    // COMMENT
+					"NULL",                // Expression
+					"",                    // INDEX_COMMENT
+				)
+				rows = append(rows, record)
+			}
+		}
+	}
+	nameToCol := make(map[string]*model.ColumnInfo, len(table.Columns))
+	for _, c := range table.Columns {
+		nameToCol[c.Name.L] = c
+	}
+	for _, index := range table.Indices {
+		nonUnique := "1"
+		if index.Unique {
+			nonUnique = "0"
+		}
+		for i, key := range index.Columns {
+			col := nameToCol[key.Name.L]
+			nullable := "YES"
+			if mysql.HasNotNullFlag(col.Flag) {
+				nullable = ""
+			}
+			colName := col.Name.O
+			expression := "NULL"
+			tblCol := table.Columns[col.Offset]
+			if tblCol.Hidden {
+				colName = "NULL"
+				expression = fmt.Sprintf("(%s)", tblCol.GeneratedExprString)
+			}
+			record := types.MakeDatums(
+				infoschema.CatalogVal, // TABLE_CATALOG
+				schema.Name.O,         // TABLE_SCHEMA
+				table.Name.O,          // TABLE_NAME
+				nonUnique,             // NON_UNIQUE
+				schema.Name.O,         // INDEX_SCHEMA
+				index.Name.O,          // INDEX_NAME
+				i+1,                   // SEQ_IN_INDEX
+				colName,               // COLUMN_NAME
+				"A",                   // COLLATION
+				0,                     // CARDINALITY
+				nil,                   // SUB_PART
+				nil,                   // PACKED
+				nullable,              // NULLABLE
+				"BTREE",               // INDEX_TYPE
+				"",                    // COMMENT
+				expression,            // Expression
+				"",                    // INDEX_COMMENT
+			)
+			rows = append(rows, record)
+		}
 	}
 	return rows
 }
@@ -384,6 +486,122 @@ func keyColumnUsageInTable(schema *model.DBInfo, table *model.TableInfo) [][]typ
 			)
 			rows = append(rows, record)
 		}
+	}
+	return rows
+}
+
+func dataForTikVRegionPeers(ctx sessionctx.Context) (records [][]types.Datum, err error) {
+	tikvStore, ok := ctx.GetStore().(tikv.Storage)
+	if !ok {
+		return nil, errors.New("Information about TiKV region status can be gotten only when the storage is TiKV")
+	}
+	tikvHelper := &helper.Helper{
+		Store:       tikvStore,
+		RegionCache: tikvStore.GetRegionCache(),
+	}
+	regionsInfo, err := tikvHelper.GetRegionsInfo()
+	if err != nil {
+		return nil, err
+	}
+	for _, region := range regionsInfo.Regions {
+		rows := newTiKVRegionPeersCols(&region)
+		records = append(records, rows...)
+	}
+	return records, nil
+}
+
+func newTiKVRegionPeersCols(region *helper.RegionInfo) [][]types.Datum {
+	records := make([][]types.Datum, 0, len(region.Peers))
+	pendingPeerIDSet := set.NewInt64Set()
+	for _, peer := range region.PendingPeers {
+		pendingPeerIDSet.Insert(peer.ID)
+	}
+	downPeerMap := make(map[int64]int64, len(region.DownPeers))
+	for _, peerStat := range region.DownPeers {
+		downPeerMap[peerStat.ID] = peerStat.DownSec
+	}
+	for _, peer := range region.Peers {
+		row := make([]types.Datum, len(infoschema.TableTiKVRegionPeersCols))
+		row[0].SetInt64(region.ID)
+		row[1].SetInt64(peer.ID)
+		row[2].SetInt64(peer.StoreID)
+		if peer.IsLearner {
+			row[3].SetInt64(1)
+		} else {
+			row[3].SetInt64(0)
+		}
+		if peer.ID == region.Leader.ID {
+			row[4].SetInt64(1)
+		} else {
+			row[4].SetInt64(0)
+		}
+		if pendingPeerIDSet.Exist(peer.ID) {
+			row[5].SetString(pendingPeer, mysql.DefaultCollationName)
+		} else if downSec, ok := downPeerMap[peer.ID]; ok {
+			row[5].SetString(downPeer, mysql.DefaultCollationName)
+			row[6].SetInt64(downSec)
+		} else {
+			row[5].SetString(normalPeer, mysql.DefaultCollationName)
+		}
+		records = append(records, row)
+	}
+	return records
+}
+
+const (
+	normalPeer  = "NORMAL"
+	pendingPeer = "PENDING"
+	downPeer    = "DOWN"
+)
+
+func dataForTiDBHotRegions(ctx sessionctx.Context) (records [][]types.Datum, err error) {
+	tikvStore, ok := ctx.GetStore().(tikv.Storage)
+	if !ok {
+		return nil, errors.New("Information about hot region can be gotten only when the storage is TiKV")
+	}
+	allSchemas := ctx.GetSessionVars().TxnCtx.InfoSchema.(infoschema.InfoSchema).AllSchemas()
+	tikvHelper := &helper.Helper{
+		Store:       tikvStore,
+		RegionCache: tikvStore.GetRegionCache(),
+	}
+	metrics, err := tikvHelper.ScrapeHotInfo(pdapi.HotRead, allSchemas)
+	if err != nil {
+		return nil, err
+	}
+	records = append(records, dataForHotRegionByMetrics(metrics, "read")...)
+	metrics, err = tikvHelper.ScrapeHotInfo(pdapi.HotWrite, allSchemas)
+	if err != nil {
+		return nil, err
+	}
+	records = append(records, dataForHotRegionByMetrics(metrics, "write")...)
+	return records, nil
+}
+
+func dataForHotRegionByMetrics(metrics []helper.HotTableIndex, tp string) [][]types.Datum {
+	rows := make([][]types.Datum, 0, len(metrics))
+	for _, tblIndex := range metrics {
+		row := make([]types.Datum, len(infoschema.TableTiDBHotRegionsCols))
+		if tblIndex.IndexName != "" {
+			row[1].SetInt64(tblIndex.IndexID)
+			row[4].SetString(tblIndex.IndexName, mysql.DefaultCollationName)
+		} else {
+			row[1].SetNull()
+			row[4].SetNull()
+		}
+		row[0].SetInt64(tblIndex.TableID)
+		row[2].SetString(tblIndex.DbName, mysql.DefaultCollationName)
+		row[3].SetString(tblIndex.TableName, mysql.DefaultCollationName)
+		row[5].SetUint64(tblIndex.RegionID)
+		row[6].SetString(tp, mysql.DefaultCollationName)
+		if tblIndex.RegionMetric == nil {
+			row[7].SetNull()
+			row[8].SetNull()
+		} else {
+			row[7].SetInt64(int64(tblIndex.RegionMetric.MaxHotDegree))
+			row[8].SetInt64(int64(tblIndex.RegionMetric.Count))
+		}
+		row[9].SetUint64(tblIndex.RegionMetric.FlowBytes)
+		rows = append(rows, row)
 	}
 	return rows
 }
